fix(js): added requred args param to cordova.exec calls
Args is a required, final parameter to the `cordova.exec()` call. When absent in recent versions of cordova, the app crashes because `cordova.exec` tries to loop through the args array but JavaScript "Cannot read property 'length of undefined".

I added an empty array to all `exec` calls wherever the args param was omitted.