feat(all): Add build process, release browser platform as a UMD library
The new build process makes it possible to bundle the entire browser platform implementation as a
library which can be used without cordova-browser. This is particularly valuable for projects which
bundle apps with Electron or NW.js and do not use the Cordova Browser platform, but still need an
API compatible implementation for desktop. The build process also makes it possible for the Web
Worker code to be bundled with the plugin code, the adapter.js and qrcode-reader dependencies to be
fully managed by NPM, and all released code to be minified. This change also adds a simple manual
test for the library and some documentation.

Fixes #30, removes some unused code in the browser implementation.